Created LinkedList Datastructure
The linked list datastructure created with a few test cases. 

My goal was to have a better understanding of how to create and use the linked list data structure. 

There are a few benefits to using this structure, it's quick and easy to implement and it's easy to add/remove data types. 

Downside is that it's difficult to traverse the list because you have to go through each individual element.